API change - AppGrid2: Separated into GridPanel and EditorGridPanel versions
AppGrid2 is now back to a normal (non-Editor) Grid. GridPanel and
EditorGridPanel are almost identical API-wise, but there are some
important differences with row selection code.

Created new xtype 'appgrid2ed' to be used when EditorGridPanel is
needed. On the Perl side, in order to have an EditorGridPanel in a
RapidApp::AppGrid2 class, the xtype must now be set in the BUILD method:

 $self->apply_extconfig( xtype => 'appgrid2ed' );