Added RapidApp::DBIC::RelationTreeFlattener
This takes a RelationTreeSpec and uses it to define a mapping from DBIC treed values to a single-level
hash suitable for ExtJS Stores.  It can be configured so that unexpected keys are either ignored, or
throw an exception.

By using one of these for reads, and another for writes, we can enforce basic permissions on the web client.