DbicLink: --- RECORD UPDATE SUPPORT ---
In this commit, for the first time, update support is built into
RapidApp::Role::DbicLink as a generalized feature with a light/simple
API provided to consuming Modules. Previously, only read support was
handled this way, and update support required creating a custom
"update_records" method (which is an API feature of DataStore2, not
DbicLink).

The main thing that was holding this feature back was the logic required
to process a flat record packet from ExtJS and separate it back into its
original separate tables, and identify which tables are even allowed to
be updated at all. This is the logic that has been added in this commit.

New API:

There are 2 new API options (Moose attributes) that must be set in order
to make use of this new capability:

 * dbiclink_updatable (Bool, default 0)
 * dbiclink_updatable_relationships (ArrayRef, default [])

If dbiclink_updatable is set to 1 (true) then an update_records method
is dynamically added to the Moose object which will handle a standard
data packet from the update call of Ext.data.JsonStore and translate it
into DBIC updates, called within a transaction. By default, only the
columns of the base row are updated, and the others that were the result
of joins are discarded.

However, related rows included in dbiclink_updatable_relationships will
be updated as well, as long as they are single (return a Row) and not
multi (return a ResultSet).

TODO:

Currently the update_records method just returns 1 upon success. We want
to make this return a complete ExtJS Store return packet with the new
flattened values after the update. This will require a bit more logic,
however, as the context of an update is different from a (original read,
which has quite a bit of logic to do the flattening in the first place)

MORE INFO:

This is discussed in more detail in WebIRS Issue 12815