BaseRs: Added 'native_rs' back to RapidApp::DBIC::ResultSet::BaseRs
Had to use a fairly exotic approach to get this to work. See the
comments in the file for details.

Also added a modified version of the CPAN module Data::Dmap as
RapidApp::Data::Dmap. The latter is a hacked version that supports perl
5.8 (the CPAN module requires 5.10). This just involved back porting to
the "Switch" module since 5.8 has no native switch statement.