Role::DataStore2: Fixed bug in 'defer_to_store_module'
Now use 'getStore_code' instead of 'getStore' - getStore_code is
javascript code that gets the store. This is needed, otherwise the store
gets overwritten in the client side which causes it to be bound only to
the last component that shares the store. Now it is properly bound to
all components that use it