Started creating RapidApp::TableSpec
This class is a generalized configuration/specification object that can
hold config info about a given data/table structure for use in various
modules that may use such information. The purpose is to consolidate all
this information in one place that can then be consumed by Grids, forms,
etc.

Also added a new DBIC Component class that will automatically attach a
TableSpec object with a Result class, and pre-configure it with
information available from the Result class.