DataStore2: Preliminary config loading via TableSpec
DataStore2 based modules now load column definitions from 'TableSpec' if
defined/supplied.

Also added logic to support adding ordered columns with an arbitrarily
high array index offset. If the offset is higher than the length of the
array (number of columns) then it is just set to the end of the array.