DbicLink: Result class TableSpec support
DbicLink now checks the Result class associated with the ResultSource
for a 'TableSpec' (as setup by RapidApp::DBIC::Component::TableSpec) and
if it is found it is automatically configured as the TableSpec attribute
for the module.

This can optionally be disabled by setting the new attribute
'ignore_Result_TableSpec' to true. If another TableSpec is manually
defined, it will take precidence.