RapidApp::TableSpec: Refactored and changed API
New API requires both of these lines to be added to a Result Class:

__PACKAGE__->load_components('+RapidApp::DBIC::Component::TableSpec');
__PACKAGE__->apply_TableSpec;

This is the proper way to build Class::C3 componenets. The
load_componenets call isn't meant to do anything but make new methods
and accessors, which then subsequently need to be called. It is working
much more reliably now, and is working correctly with multiple Result
Classes at once.

Also built working code to be able to pull in columns from the TableSpec
of a related Class. This functionality still needs the kinks worked out.