Added @RapidApp::RapidApp::GLOBAL_INIT_CODEREFS
This provides a mechanism to queue coderefs to be called after the
RapidApp root module has been loaded. This will be used in the TableSpec
code that I am trying to get to setup AppCombo2 modules automatically.
Since this TableSpec code is called from the DBIC Result classes, which
get loaded well before the RapidApp modules, this feature is needed.