Added 'conflictResolver' feature to RelationTreeFlattener
In order to prevent time-consuming exceptions on each conflicking column, I added
a callback that can be specified during the RelationTreeFlattener constructor
to resolve conflicts on the fly.

I also changed RelationTreeSpec to look for columns first, before relations,
for the pathological case of a ResultSource having a column and relation by
the same name.