Changes to @GLOBAL_INIT_CODEREFS and Functions
Moved @GLOBAL_INIT_CODEREFS execution from happening after the Root
Module is loaded to the start of the Root Module loading process: after
the Root Module itself has been loaded but before any of the other
modules in the application have been (in BUILD in RapidApp::RootModule).
This was needed so that code can take effect after the RapidApp
environment is setup but before the RapidApp modules are loaded.

Also fixed scream_color in RapidApp::Functions