TableSpec: Working 'TableSpec_add_relationship_dropdown_columns'
Initial functional setup of relationships as virtual columns with
associated combo editor. The column properties passed to this setup
function take the nomral column properties, but also special properties
that are used to setup the combo and rendering:

 * displayField (required)
 * valueField
 * render_col

Typically only displayField is needed. Here is an example where
'creator' is a defined 'belongs_to' relationship (i.e. single - multi
relationships don't make sense in this context):

This would be added to the Result class after loading the
RapidApp::DBIC::Component::TableSpec component:

__PACKAGE__->TableSpec_add_relationship_dropdown_columns(
  creator => {
    displayField => 'full_name',
	 header => 'Creator',
	 width => 120
  }
);