AppDV: Started 'autofield' - support for rederer funcs
Started on new field type 'autofield' (along side edit_field and
edit_big_field). The purpose of this will be to detect and setup field
params automatically, such as editable and render params.

If the field/column has a renderer, it will now be used to display the
field using autofield. Editable logic to follow...