DataStore2: Added onrequest_columns_munger option
If defined, the DataStore2 columns hash will be passed to the
RapidApp::Handler object in the onrequest_columns_munger attribute.
Because 'columns' is a RapidApp per request attribute, it can be changed
within the munger and will be reset on the next request. This provides a
mechanism to change column properties based on the current user (such as
removing editors to enforce permisions)

AppPropertyGrid: editable_fields is now automatically set based on the
existence of an editor. Only columns with editors (or passed in the
editable_fields hash) are editable in the AppPropertyGrid. This is a
difference from the normal Property Grid which has all fields editable
per default