TableSpec/DbicAppCombo2: Added 'valueqry' support to DbicAppCombo2
Expanded the DbicAppCombo2 API to operate in two modes: normal and
valueqry. In valueqry mode it is looking up a single, existing value,
whereas in normal mode it is pulling up the dropdown list, which may be
filtered. By supporting the alternate mode, existing values that cannot
be selected from the list will still display properly.

TODO: This needs to be brought into the fold with DbicLink. Not doing
that yet because I am going to allow actual feature needs drive how that
API should look and be. For now, DbicAppCombo2 is simple enough that
there is no need to do this now.