HEAVILY optimized the Column class, maybe more than necessary.
I rewrote RapidApp::Column to not be based on Moose.
I used a very procedural and eval-heavy approach which is
unpleasant to read, but should give the absolute fastest
performance.

It shaved 46 seconds (>50%) off startup time, for me, so it
might be worth the price.

This isn't really saying anything bad about Moose, just that
we're creating 114355 objects, and Moose isn't designed for
that scale.