Greatly improved the process interaction between background tasks.
BgTask now has new behavior for the Supervisor process:
  * It no longer becomes its own process group
  * It forks/exits one additional time so that its parent becomes
     'init', and the real parent doesn't have to watch for it
     to exit (because there was no good way to do that, and we
     don't want defunct processes hanging around)
  * It sets a timer (6 sec) that watches for the leader of its
     process group to exit.  If the group leader exits, the
     script exits as well (terminating its job).
  * If we decide that we want it to outlive its group leader,
     we can pass the option "behavior => 'daemon'" to spawn(),
     and the supervisor will become its own session.
     (fully detached).

During TaskPool->spawn, there were a few places where the process had been created,
but it failed to perform some phase of the startup sequence, and the method exits
with an error.  I had forgotten to call waitpid on the dead process in these cases.
I also call waitpid to reap the temporary process created by "background mode".

I also added a try/catch around the code that tries to connect to a process and
read its metadata, which resolves any case where a hung or not-reaped process
would cause the connect to fail.