Bug and documentation fixes in Supervisor.
Was using kill to determine existence of the process group leader, but this only
works when running as root or when group leader is same user as Supervisor.
Changed it to test for existence of /proc/$PID

Updated the debugging instructions at the top of the file.

Stdout was getting written to an ".out" file with the wrong PID name.
I delayed the creation of the output file until the final fork is done.