DbicLink2: Handling of no updatable/creatable colspecs
At the DbicLink2 layer, the default values of updatable and creatable
colspec is undef, and in that case they are not passed to the TableSpec
object. This is also now checked to decide whether or not to add the
update_records and create_records methods (i.e. if no update_colspec is
supplied, then there will be no update_records setup)