AppGrid2: Added 'store_buttons' feature
AppGrid2 can now be created with an array param "store_buttons" which
turn on buttons, next to reload in the paging toolbar, to perform basic
CRUD operations. The names of these buttons (identified by name/string)
are:

 * add
 * delete
 * reload
 * save
 * undo

The preconfigured buttons get created and added to the toolbar to
perform associated actions on the store. These buttons are fully
configured along with listeners to enable and disable them based on if
there is a record selected (for delete) and if there are outstanding
changes (for save/undo). Additionally, 'add' is ignored if there is no
"create" api and 'delete' is ignored if there is no "destroy" api