DbicLink2: Added 'get_ResultSet' option
New CodeRef attribute to return the *base* ResultSet. Default returns
'$self->ResultSource->resultset'. This provides another layer of control
beyond declaring the "ResultSet" method, in that, as an attribute,
"get_ResultSet" can be passed to the constructor. What is returned by
get_ResultSet is what is passed to the "ResultSet" method as the first
argument, if defined.