AppGrid2 + DataStore2: autoLoad and store_columns
DbicLink2: DataStore2 'store_columns' are now set per default. Without
this setting DataStore2 dynamically figures out the columns based on the
data it gets back, but this doesn't work for the very first record on
create. Previously in all of these cases store_fields were manually set
in DataStore_build_params, but now its built-in/automatic. The fields
are set to all of the columns from the TableSpec plus record_pk and
'loadContentCnf' which is specifically for AppGrid2.

store_autoLoad is also turned on per default as well, and this is now
setup as an object (hash) instead of a bool in DataStore2. This hash is
set with start and limit values, which in turn get passed to the first
load call from the store.

AppGrid2 now uses autoLoad instead of expressly calling load (now that
autoLoad is default, this prevents 2 load calls). autoLoad is also
modified, setting the "limit" param to match the page size.

This change also fixes the long standing issue where the load mask
wouldn't get displayed on the first load.