AppGrid: Built-in support for 'autoExpandColumn' feature
The "id" property is now automatically set on the column
autoExpandColumn refers to. It is set to teh same as the name. Don't
know why this feature matches on id (while others match on name, and
others on dataIndex). For our purposes, we want them to just be the
same as name.