AppTab: Tab title/icon can now be set from the loaded content
The 'autopanel' loaded content of a tab now supports the following
optional properties:

 * tabTitle
 * tabIconCls

If set, after the content has been loaded into the tab via autopanel the
parent tab's title and iconCls properties are set to these values.
Additionally, if no title/icon is supplied to the loadContent call, a
temporary "Loading" title and loading icon (animated gif) are set until
the content is loaded, and if it doesn't specifiy the title/icon
"Unnamed Tab" and a generic "page" icon are set. title/iconCls can still
be supplied to loadContent and if the loaded content lacks the new
params it will behave as it always has...