DbicLink2: single_record_fetch attr + new count func for multi-rel cols
New boolean Moose attribute 'single_record_fetch' that when set to true
in DbicLink2 modules will only fetch 1 record from the database and will
NOT call the second query for the total count. This is useful on
property pages that load only a single database record

Also enhanced the logic for the multi-relationship columns count
function. This is now working a lot better but is still not the final
design (planning to try calling separate queries for each in a
munger/around type design