datastore-plus: 'use_add_form' supports Tab or Window
The recently added 'use_add_form' can now optionally be set to 'tab' to
load the form in a Tab in the tabpanel found by global id
'main-load-target'. The default is Window, which is used if use_add_form
is any true value other than 'tab' (i.e. 'window')

Also made improvements and cleanup to the code in general