datastore-plus: added optional param 'autoload_added_record'
If autoload_added_record is true then records added will try to be
loaded using the 'loadContentCnf' in the record - just like a
double-click from appgrid2.

Did a minor bit of refactoring for this. Will do a full overhaul soon...
for now it is a simple drop-in feature to the existing nav stuff