DataStore2: ADDED 'delete_columns'
DataStore2 now has a working delete_columns method. This allows
deleteing columns in the context of the current request (because of the
per request attr stuff) and mainly is useful for doing "personalities"
where a module can have a different set of effective columns for
different requests/objects but still be the same module.

Using this in DbicAppPropertyPage to remove extra columns (performance
only)

Also added corresponding DbicLink2 sugar methods: delete_colspec_columns
and delete_except_colspec_columns

NOTE: in order for delete_columns to operate in a realistic/sane way, it
does not allow columns to be added back again once they have been
deleted (via apply_columns). But again, this is only the case until the
end of the request. See code comments for more details