MultiFilters/QuickSearch: SUPPORT TableSpec relationship columns
MultiFilters and QuickSearch have now been updated to support all the
new tablespec stuff that has been added, specifically relationship
columns.

Added 2 new column params to support this:

 * query_id_use_column
 * query_search_use_column

If defined, MultiFilter will swap the column name for these values.
query_id_use_column is used for 'is' when the supplied editor is used.
This does a direct lookup, it is set as the "real" key value for
relationship column dropdowns.

With any other cond other than 'is', query_search_use_column is used
instead. with relationship columns this is set to the render column, and
it is what users see as the value of the field. This oen is used so they
can do string searches, typed in directly

The quick search swaps for query_search_use_column too

Also did some other tweaks and changes