DbicLink2: read_records: distinct=>1 - TEMP DBIC BUG WORKAROUND
Turned on 'distinct' => 1 across te board (read_records). This has not
been turned on previously because it would break the ability to get the
total count. As it turns out this is a DBIC bug. Implemented a work
around - see in-code comments.

This needs to be fixed as soon as DBIC is fixed