TableSpec: Support for multi-relationship attrs
multi relationships (i.e. 'has_many') can be defined with attrs,
including join and where, and those conditions are honored by the
multi-rel system for calculating correct counts, as well as loading the
grid for those.

This is important for the ability to define multi-rels with arbitrary
conditions, such as a relationship to "object" but specific to a given
object type (DataRole/object schema designs). This makes the feature
totally flexible.

This is still very ugly and preliminary and has code specific to
GreenSheet (setting "personality". Needs to be refactored and cleaned
up.