DataStore2: Moved add_form logic into the backend
Moved the logic that gets the add record form out of the datastore-plus
JS plugin and into RapidApp::Role::DataStore2. The form is now fetched
via Ajax when the add button is clicked.

This is done via a new action "add_form" which calls the sub
"get_add_form" which returns a HashRef config of the formpanel. This is
added automatically if the store has 'create'