datastore-plus: Add form only closes upon successful create
If the store is configured for instantly persisting create operations
the add form will not close until after the successful create round
trip. If the create fails, the form stays on the screen.

This way the user can try again without losing the form data
(assuming a useful message is returned opun failure, such as a duplicate
name, etc)