Made trace-capture optional
The TraceCapture feature installs a SIG{__DIE__} handler which collects a stack trace.
These stack traces are saves until the end of the request, where they are stored if
the request ended badly.  If the request succeeded, they are discarded.

If nothing is saving the errors, then they aren't needed, and can potentially cause
internal perl errors due to abuse of magic in DBIC.  The default now is to have them
disabled.