DataStore2: Added 'menu_select_editor' feature
This is a new special column property that sets up a renderer and editor
based on a predetermined list of values. This is an alternative to a
bool/checkbox, and can have more than 2 values, can set the render
icon and display for values, etc. Currently the editor is a static
icon combo

Example:

		menu_select_editor => {

			render_icon_only => 1,

			selections => [
				{
					iconCls => "icon-cross-light-12x12",
					text	=> 'No',
					value	=> 0
				},
				{
					iconCls => "icon-checkmark-12x12",
					text	=> 'Yes',
					value	=> 1
				}
			]
		}