ClickActionField(s): fixed bug where action is loaded in forms
Check is now done on show to see if any value has been set in the field.
If so, then it is likely a show from a grid editor (which means the user
clicked) and we should display the action (menu, image selector, etc).
If the value hasn't been set then it may be in a form and we don't want
the action displayed until the user clicks it. In this case, it is
important to set a default value with 'value' so something is displayed.