DataStore2: refactored 'delete_columns' + extended API
New approach to "delete_columns" that doesn't actually alter the
contents of $self->columns but rather applies a filter when column_list
is called.

Also extended the API with "get_column" and "has_column".
$self->get_column($col) must now be called instead of $self->columns->{$col}
in order for deleted columns to work properly. This change was made to
overcome sporadic caching issues with delete_columns when used
dynamically ONREQUEST