Added Ext.ux.RapidApp.Plugin.RelativeDateTime ('form-relative-datetime')
General-purpose plugin for DateTime fields. Allows dates to be specified
via human friendly duration string, relative to the current date.
Duration strings must start with '+' or '-' to indicate forward or
backward from the current time. Valid duration string exaples:

 * -1 year, 15 days
 * -1y15d
 * +5 months
 * -24 hours, 59 mins
 * -60secs

This plugin is now on by default in date and datetime TableSpec column
profiles, as well as in 'date' condition type in MultiFilter