MultiFilter/DbicLink2: Relative Dates in Searches work
Expanded the capabilities of the new 'form-relative-datetime' plugin,
adding an option to leave the human date string as-is instead of
converting. Turned this new option on for the multifilter date fields,
which makes them get submitted to the server. Added server-side logic to
convert them into dates.

TODO: Add an interface to the DatePicker for relative dates. Relative
dates can be typed in, but users still have to know to do this...