RELATIVE DATES: Now support "reference date" keywords
Relative dates can now be prepended with a even/threshold reference date
keyword such as "today" or "this week"

For example, if the current time was '2012-04-22 11:43':

-1 day 		# evaluates to  '2012-04-21 11:43'
today -1 day 	# evaluataes to '2012-04-21 0:00'

The keyword "today" makes the reference date/time midnight of the
current day instead of the currect exact time. Other keywords include
"this week", "this month" and "this year". The keyword "now" is also
supported, which is the default (i.e. "now -1day" is the same as
"-1day")

Also expanded the css styles of the relative date input menu to describe
this new feature. Also, the calculated date is now displayed so the user
knows what effect their input has immediately.