MultiFilters: MULTI-REL COLUMNS NOW SUPPORTED
Virtual, "multi relationship columns" can now be searched/filtered
within MultiFilters. These columns are actually sub-query, virtual
columns that perform a count. Searching on them requires using a HAVING
clause instead of a normal WHERE clause, which is now automatically
detected and setup in the back-end.

This also required some tweaks to the code that obtains the total count.
See in-code comments for more details