TableSpec Columns editor sets 'maxLength' according to schema
The editor field that is automatically setup via DBIC/TableSpec now
applies 'maxLength' based on the length/size of the underlying column in
the schema. So, a 'varchar(32)' when edited would be shown as invalid
with 33 characters or more.

This required extending the TableSpec/Column API beyond 'profiles' with
'extra_properties' to allow finer-grained control.