DbicLink2: New attr option 'natural_column_order' (default true)
Added new option 'natural_column_order' that reverts the column order to
the original order defined in the Schema class. This is now the default.
Previous behavior, which is to set the order according to the
include_colspec (which is not predictable when using wildcards),
requires setting 'natural_column_order' to 0 (false).

Note that this (typically) will have no effect when loading a saved
search because it contains column order info directly