AppGrid Paging Toolbar: many enhancements, displays query_time
Many improvements to Ext.ux.RapidApp.PagingToolbar. Page size now
displayed in the toolbar, also displays the value of 'query_time'
if it was returned in the response by the server.

Note that many of the options that are configurable in the normal Paging
Toolbar are now hardcoded. This was done to avoid the need for lots of
extra code, when the paging config is generated automatically now
anyway. This opts are now forced: beforePageText (now empty),
displayMsg, and prependButtons (forced to false)