MultiFilters: Fixed deep bug in rel column + date handling
The logic for handling multifilters for relationship columns which had
been moved to the back end and appeared to be working wasn't actually,
do to what turned out to be a simple logic flaw, but it was hidden
because the lookups were happening based on the display column, which
works too, but has issues. Now this is fixed and id values are used as
designed.

Also had to split out datefields into a special case using 'getRawValue'
while other kinds of fields must use normal 'getValue'