DataStore2: TURNED OFF 'writeAllFields' option
This option has been on forever, and for no good reason. It causes store
updates to send ALL fields. With it off (default) it only sends fields
that changed. This change doesn't appear to effect performance (even
though less fields are being sent) but it addresses situations with
concurrent changes to different fields of a record undoing eachother,
either from the same user with multiple pages/tabs open covering the
same record, or across different users.

It is a better default to have this option off (Ext's default) and this
is now set