Excel Export: Now uses newer XLSX file format + other improvements
Converted from using the Perl module 'Spreadsheet::WriteExcel' which
writes Excel 97-2003 format (XLS) to 'Excel::Writer::XLSX' which writes
Excel 2007-2010 format (XLSX). The primary reason for this change was to
support more rows than the 65536 limit of XLS spreadsheets.

The API of Excel::Writer::XLSX is compatable with
Spreadsheet::WriteExcel so this was able to mostly be a drop-in
replacement.

Also increased the hard max rows limit in DbicLink2 from 100,000 to
10,000,000.

Also changed the way cells are written to dynamically set type which
removes the problem where number cells would have the little green
triangle and say "number as text". Now numbers are numbers which allows
them to be used in Excel formulas without having to convert first.

Finally, the MIME Content-Type of the download is set specifically to
the Microsoft Excel 12 type 'application/vnd.ms-excel.12' instead of the
generic 'application/x-download'. It is the 'Content-disposition', not
the Content-Type, that makes the file show up as a download, and the
more specific type makes download dialogs correctly display that the
download is an Excel file.