DataStoreAppField: background loading/caching functionality
Instead of creating a new Window and loading the app on every display,
it is now loaded in the background and hidden. This provides
significvant performance improvements. Lots of code and logic to support
this functionality in a sane way, including new code to "take over"
store loading events, and reload the store every time the window is
re-shown