DataStoreAppField: opportunistic loading display values
Significant improvements and structure for tracking the state of a raw
value vs a display value, and opportunistically loading the display
value as soon as it is seen in the store. It even has a nice updating
indicator while the value is 'dirty' and hasn't yet been set with its
correct display value.

Next I need to add an active (vs opportunistic) query to the server to
resolve display values if the grid isn't immediately being opened. This
is important in cases where the field is being displayed as a field (not
as an editor). This will happen most of the time in MultiFilters, but
could also happen in any other form context with a pre-populated value.

None of this matters unless the field has different displayField and
valueFields (same paradigm as standad Ext combo). If they are the same
(or if one of them is not defined) this code is deactivated and raw
values are used.