AppCombo2: NEW FEATURE/OPTION 'allowSelectNone'
AppCombo2 (and classes that derive from it) has a new bool parameter
'allowSelectNone' (defaults to false). If true, a special '(None)' item
will be added to the top of the combo list which unsets the value of the
field.

TableSpec: nullable relationship columns (combos) are now setup with the
new 'allowSelectNone' turned on. This allows the value of these fields
to be set back to null. Prior to this addition, these column could be
null '(not set)' but once a value was selected there was no way to set
it back to not set. This 'Select None' feature was recently/separately
added to 'grid' combo, but now simple combos have this ability, too
(since there are many cases with a small list of values where a simple
dropdown is prefered over a full-blown grid combo and we want the field
to be nullable)