TableSpec/DBIC: VIRTUAL COLUMNS SUPPORT
Virtual columns can now be added in the TableSpec Result Class with
'add_virtual_columns' which support special attrs 'sql' or 'function' to
define any sql to use when selecting the column in queries. The 'sql'
will translate 'self.' into 'me.' or the correct join name.

TODO: There are issues with multifilters and virtual columns, but these
issues already existed and also apply to multi-rel columns. Both need to
be fixed.

See in-code comments for more info