DbicLink2/MultiFilters: MAJOR 'HAVING' vs 'WHERE' REFACTORING
Major refactoring work to clean up and fix edge cases and issues with
virtual column and MultiFilter searches. Completely refactored the
logic/flow/design for the dynamic 'HAVING' vs 'WHERE' query construction
for virtual columns.

The HAVING/WHERE functionality was a bleeding edge feature added to
allow MultiFilter searching in multi-rel columns, which are virtual and
thus not accessible via normal WHERE. The new, more general "virtual
columns" feature also needs to search using HAVING, like the earlier
multi-rel. Ran into issues and discovered complex interdependency
problems that existed with multi-rel columns but were not as noticible
because the only "value" to search in a multi rel is the count, but also
exist when trying to extend the existing design to support the general
virtual columns (which can have any value).

The fundamental problem is that since HAVING and WHERE clauses are
independant of eachother within the broader scope of the query as a
whole, when mixing HAVING conditions with WHERE conditions the complex
and sophisticated -and/-or logic built in MultiFilters becomes broken
and disjointed. I discovered this when doing tests with combinations of
critiera with ands and ors, virtual and nomral columns.

After much effort and investigation, I came to the conclusion that there
is no safe/sane way to split a MultiFilter condition into separate WHERE
and HAVING parts. The only way for it to function correctly is to either
have the *whole* condition set represrnted by a WHERE, or the whole set
represented by a HAVING.

In order to support HAVING, a whole new approach was needed that toggles
a a query-wide 'needs_having' flag, and then convert the *entire* WHERE
clause into a HAVING clause. This requires significant remapping of data
structures, multi-depth tracking of information callbacks, etc. HAVING
really only works with 'virtual' columns, so this meant that all columns
that would otherwise be simple/where conditions have to be converted
into virtualized SQL monikers (with AS, like all virtual/function
columns).

This was significant, and while this may sound ugly, the implementation
that has now been completed in this commit is actually very solid, and
the whole HAVING sub-system is now much better abstratced and pulled
into the rest of the system/logic. I've also done a lot of testing and
all combinations of exotic multifilters now seem to be working properly.

There is also a LOT of in-code documentation in comments, so refer to
that for more details and information.