BaseRs: Properly/completely fixed count "Issue 13305" problem
Discovered that BaseRs was broken by 'as_subselect_rs' and then realized
that is the root problem of the earlier work-around fix ("WebIRS Issue
13305"). Found exactly where the breakdown was occuring and fixed with
an extended as_subselect_rs method and a few other tweaks. See in-code
comments for more details.

Default total_count logic in DbicLink2 updated and enhanced using this.