DbicLink2: switched from rs_count_manual -> rs_count_with_fallbacks
After confirming that the rs_count_manual function *is* meaningfuly
slower than the DBIC count via pager mechanism, switched to use the
earlier developed rs_count_with_fallbacks function which tries the pager
first, and then falls back to rs_count_manual if that fails. I held off
turing this on earlier in order to prevent adding a lot of code
complexity if there was not a big benifit. Now that I have run more
tests and verified it does provide a meaningful benifit, it has now been
turned on. Also, this *does* add exception code into the anticipated
codepath, which is not ideal, but worth it because of the now known
performance differences.

Background: The reason the rs_count_manual function was added in the
first place was that there are known cases where the pager will fail
with virtual columns, and the rs_count_manual function is reliable/safe.
This is a very small number of places, and so now it is only used when
it has to be used since it is slower.