IN-PROGRESS: REFACTORED BROWSER HISTORY - new HashNav design
The browser history and url path '#' is now directly tied to the
url/params path of the LoadContent of the tab. When opening a new tab,
the # url is set with the 'hashpath', and conversely, when opening a
hashpath url, the associated tab is activated, or if it doesn't exist,
is opened new. This makes the browser history tied to the
real/meaningful content paths, and even supports direct navigation to a
tab via absolute URL from the outside (if logged in)

This system is functional now, however, other components and modules
need updating to be fully compatable...