AppCmp: optional stash-based appliation of extconfig params
New mechanism now available to for setting extconfig params via the
Catalyst stash object. This was added to allow normal catalyst
controllers to be able to dispatch to RapidApp Modules and be able to
configure fine-grained extconfig params. This is the same concept as
calling 'apply_extconfig' from within the RapidApp Module at various
stages of the request cycle.

This was specifically added for new dynamic 'RequestMapper' controller,
added to greensheet, to map friendly url paths to complex loadcontent
params, such as loading saved searches