REST QUERY KEY NOW OPTIONAL + misc fixes
When opening a content/tab via DbicLink2 rest query/path, the key is now
optional, and if not specified, the default key name is assumed. For
example, assuming the default record_pk value of '___record_pk', the
following are now equivalent:

 * somesite.com/sometable/item/___record_pk/1234
 * somesite.com/sometable/item/1234

The latter is now the default path opened by double-clicking in
AppGrid2. Since the key can now be hidden, the rest alias for
___record_pk (which had been 'id') is no longer a needed feature and has
been removed