TableSpec/DBIC: refactored single rel renderer (REST link)
Revamped the renderer for single relationship columns to use a simple
REST link as long as the relationship references the primary ket and
there is a TableSpec 'open' link. Otherwise, it falls back to the older,
more complicated loadCfg inlineLink.

Now the REST links generated in the renderer should should exactly the
ones in the AppGrid row nav, $Row->inlineNavLink, etc