X-RapidApp-EVAL header + global var 'Ext.ux.RapidApp.NO_DBIC_REL_LINKS'
If the global variable Ext.ux.RapidApp.NO_DBIC_REL_LINKS is true, open
links are not shown with relationship columns. Added check for response
header 'X-RapidApp-EVAL' which provides a mechanism to eval code. THis
was added to provide a mechanism for changing the value of
NO_DBIC_REL_LINKS programatically from the back-end, but it could be
used for anything.

This design is temporary and should be refactored later into something
not so specific and potentially dangerous