AppGrid2: New feature/config attribute: 'open_record_column'
New moose/config param bool attribute 'open_record_column' (defaults to
false) if set to true causes an extra, special open action column to be
added to the grid that can be clicked on to open the record, just like
double-clicking. The link is a magnify glass icon with a tooltip "open
record". This is useful for clients that cannot double-click, such as
iPads.

The column header also shows a tiny magnifying glass icon, with the text
"(open link column)" visible in the columns menu for toggling the column
visible/hidden. Only the tiny magnify icon is visible in the actual
column heading when active (fixed width of 30 px)

Config param 'open_record_column_init_visible' (also defaults to false)
is also now available which determines if the open_record_column should
be automatically visible, or hidden (default).

Note: The column isn't a "real" column and thus is currently
ignored/excluded from Saved Searches. TODO: add this support