DataStore2: integrated before close save & confirm delete dialogs
Integrated the existing "Unsaved Changes" on close dialog with the new
'Confirm Delete' dialog. The Unsaved Changes dialog, which is shown if a
page/container is closed with pending changes, specifically warns the
user if the pending changes include deletes, and overrides the confirm
delete dialog from being displayed as well. This fixes a bug which would
cause inconsistences when both dialogs would be displayed in this
scenario