TableSpec/DBIC: WRITABLE VIRTUAL COLUMNS SUPPORT
Virtual columns now support being writable (insert and update) by
defining the column attribute 'set_function'. This should be a coderef,
which will be called with the Row object and new column value supplied
as arguments. Then any arbitrary code can be executed for the update.

Here is an example of a virtual column that is a CSV string which maps
to a has_many relationship (aka joining into a string). The read side is
handled by the custom 'sql' function (SQL code), and the write side is
handled by the custom 'set_function' coderef (Perl code):

__PACKAGE__->add_virtual_columns(
 specialnumber_list => {
  data_type => "varchar",
  is_nullable => 0,
  size => 255,
  sql => 'SELECT IFNULL(GROUP_CONCAT(customer_specialnumber.specialnumber),"") ' .
    ' FROM `customer_specialnumber` ' .
    ' WHERE customer_specialnumber.customerNumber = self.customerNumber',
  set_function => sub {
   my $self = shift;
   my $value = shift;

   $self->delete_related('customer_specialnumbers');

   return unless (defined $value and $value ne '');

   my @vals = split(/\s*,\s*/,$value);

   $self->add_to_customer_specialnumbers({ specialnumber => $_ }) for (@vals);
  }
 },
);